Youtube API
+ 확보한 카테고리 자동 입력 크롤러 구축
+ 추가적인 보수 / 작업
+ 통계수치 api 기본 틀 작성

+++ back/src/main/java/com/my/spring/domain/ChannelDto.java
@@ -1,8 +1,7 @@
 package com.my.spring.domain;
 
+import java.math.BigInteger;
 import java.sql.Date;
-
-import com.google.api.client.util.DateTime;
 
 import lombok.*;
 
@@ -14,8 +13,8 @@ public class ChannelDto {
 	private String description;
 	private String thumbnail;
 	private Date publishedDate;
-	private int viewCount;
-	private int videoCount;
-	private int subsCount;
+	private BigInteger viewCount;
+	private BigInteger videoCount;
+	private BigInteger subsCount;
 	private Date infoDate;
 }
